requirements.txt
tensorflow==2.17.1
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
datasets>=2.15.0
wandb>=0.15.0
Pillow>=9.4.0
scikit-learn>=1.2.0
